Split image builder and cluster stack generation tests
Move dummy model classes into new tests modules.
